Added custom function to make it fullscreen.
Currently the togglefullscr function tells the application to fullscreen,
which in firefox hides the tabs, bookmarks, etc.
(https://dwm.suckless.org/patches/actualfullscreen/)

So I made togglefullscr_2 (creative name, I know) to only put the window
on top, and make it fill the screen, so the window can act like it's not
fullscreened, but it actually is.